requirements.txt
Flask==3.1.2
Flask-Limiter==4.1.1
python-dotenv==1.2.1
Werkzeug==3.1.5
qrcode==8.2
pillow==12.1.0
reportlab==4.4.9
sendgrid==6.12.5
gunicorn==21.2.0
